Выполнен функционал public events

+++ main-service/src/main/java/ru/practicum/service/category/CategoryServiceImpl.java
@@ -9,28 +9,35 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.category.NewCategoryDto;
+import ru.practicum.exception.ConflictException;
 import ru.practicum.exception.NotFoundException;
 import ru.practicum.mapper.category.CategoryMapper;
 import ru.practicum.repository.CategoryRepository;
 import ru.practicum.model.Category;
+import ru.practicum.repository.EventRepository;
 
 import java.util.List;
+import java.util.Objects;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class CategoryServiceImpl implements CategoryService {
 
-    final CategoryRepository categoryRepository;
-    final CategoryMapper categoryMapper;
+    CategoryRepository categoryRepository;
+    EventRepository eventRepository;
+    CategoryMapper categoryMapper;
 
     static final String CATEGORY_NOT_FOUND = "Категория с id = %d не найдена";
 
     @Transactional
     @Override
     public CategoryDto create(NewCategoryDto newCategoryDto) {
+        if (categoryRepository.existsByName(newCategoryDto.getName())) {
+            throw new ConflictException("Категория с таким именем уже существует", "");
+        }
         Category category = categoryMapper.dtoToCategory(newCategoryDto);
         Category savedCategory = categoryRepository.save(category);
         return categoryMapper.toCategoryDto(savedCategory);
@@ -40,13 +47,24 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void delete(Long id) {
         validateId(id);
+        // Проверка на связанные события
+        boolean hasEvents = eventRepository.existsByCategoryId(id); // Метод проверки в репозитории
+        if (hasEvents) {
+            log.error("Удаление невозможно: категория с id = {} связана с событиями.", id);
+            throw new ConflictException("Удаление невозможно", "C категорией связано одно или несколько событий.");
+        }
         categoryRepository.deleteById(id);
+        log.info("Категория с id = {} успешно удалена.", id);
     }
 
     @Override
     public CategoryDto update(Long id, NewCategoryDto newCategoryDto) {
         Category existCategory = categoryRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException(CATEGORY_NOT_FOUND, "Объект не найден"));
+        String newName = newCategoryDto.getName();
+        if (!Objects.equals(newName, existCategory.getName()) && categoryRepository.existsByName(newCategoryDto.getName())) {
+            throw new ConflictException("Категория с таким именем уже существует", "Ошибка обновления категории");
+        }
         existCategory.setName(newCategoryDto.getName());
 
         Category updateCategory = categoryRepository.save(existCategory);
